Refactored all OPDS Controller methods to an OpdsService for future unit testing.
