+++ API/DTOs/OPDS/Internal/FeedAuthor.cs
