+++ API/DTOs/OPDS/Internal/FeedCategory.cs
