+++ API/DTOs/OPDS/Internal/FeedLinkRelation.cs
