+++ API/DTOs/OPDS/Internal/FeedLinkType.cs
