+++ API/DTOs/OPDS/Internal/OpenSearchDescription.cs
