+++ API/DTOs/OPDS/Internal/SearchLink.cs
