+++ API/DTOs/OPDS/Requests/IOpdsPagination.cs
@@ -1,0 +1,6 @@
+﻿namespace API.DTOs.OPDS.Requests;
+
+public interface IOpdsPagination
+{
+    public int PageNumber { get; init; }
+}
