+++ API/DTOs/OPDS/Requests/IOpdsRequest.cs
@@ -1,0 +1,9 @@
+﻿namespace API.DTOs.OPDS.Requests;
+
+public interface IOpdsRequest
+{
+    public string ApiKey { get; init; }
+    public string Prefix { get; init; }
+    public string BaseUrl { get; init; }
+    public int UserId { get; init; }
+}
