+++ API/DTOs/OPDS/Requests/OpdsSmartFilterCatalogueRequest.cs
@@ -1,0 +1,14 @@
+﻿namespace API.DTOs.OPDS.Requests;
+
+/// <summary>
+/// A generic Catalogue request for a specific Entity
+/// </summary>
+public sealed record OpdsPaginatedCatalogueRequest : IOpdsRequest, IOpdsPagination
+{
+    public string ApiKey { get; init; }
+    public string Prefix { get; init; }
+    public string BaseUrl { get; init; }
+    public int UserId { get; init; }
+
+    public int PageNumber { get; init; }
+}
