+++ API/DTOs/OPDS/Requests/OpdsSmartFilterRequest.cs
@@ -1,0 +1,12 @@
+﻿namespace API.DTOs.OPDS.Requests;
+
+public sealed record OpdsItemsFromEntityIdRequest : IOpdsRequest, IOpdsPagination
+{
+    public string ApiKey { get; init; }
+    public string Prefix { get; init; }
+    public string BaseUrl { get; init; }
+    public int UserId { get; init; }
+
+    public int EntityId { get; init; }
+    public int PageNumber { get; init; } = 0;
+}
